重新上传，修复不能支付问题

+++ doAlipay/Util/RSADataSigner.m
@@ -29,7 +29,7 @@
 
 - (NSString *)formatPrivateKey:(NSString *)privateKey {
     const char *pstr = [privateKey UTF8String];
-    NSUInteger len = [privateKey length];
+    int len = [privateKey length];
     NSMutableString *result = [NSMutableString string];
     [result appendString:@"-----BEGIN PRIVATE KEY-----\n"];
     int index = 0;
@@ -71,10 +71,10 @@
 	[formatKey writeToFile:path atomically:YES encoding:NSUTF8StringEncoding error:nil];
 	
 	const char *message = [string cStringUsingEncoding:NSUTF8StringEncoding];
-    NSUInteger messageLength = strlen(message);
+    int messageLength = strlen(message);
     unsigned char *sig = (unsigned char *)malloc(256);
 	unsigned int sig_len;
-    NSUInteger ret = rsa_sign_with_private_key_pem((char *)message, messageLength, sig, &sig_len, (char *)[path UTF8String]);
+    int ret = rsa_sign_with_private_key_pem((char *)message, messageLength, sig, &sig_len, (char *)[path UTF8String]);
 	//签名成功,需要给签名字符串base64编码和UrlEncode,该两个方法也可以根据情况替换为自己函数
     if (ret == 1) {
         NSString * base64String = base64StringFromData([NSData dataWithBytes:sig length:sig_len]);
